circuito_config/management/commands/removeallcrons.py
from django.core.management.base import BaseCommand, CommandError
import os
from crontab import CronTab

class Command(BaseCommand):
    help = 'Cron testing'

    def add_arguments(self, parser):
        pass

    def handle(self, *args, **options):
        #init cron
        cron = CronTab(user='robertopereira')

        #remove all jobs
        cron.remove_all()

        cron.write()