requirements.txt
Django==2.2.6
django-crontab==0.7.1
minimalmodbus==1.0.2
pyserial==3.4
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
